requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pypdf>=3.17.0
python-dotenv>=1.0.0

